Introduced two new configuration options for jbake.properties: - asciidoctor.attributes.export - asciidoctor.attributes.export.prefix
If `asciidoctor.attributes.export` is set to true, then all configuration key-value pairs that do not start with "asciidoctor" are appended to the asciidoctor.attributes value. Each key will be prefixed with the content of `asciidoctor.attributes.export.prefix` or "" if not defined.

Each occurence of a dot (`.`) in an existing key will be replaced by an underscore (`_`) before the key is exported as asciidoctor attribute.