Prevent image source path from breaking
This change prevents image source path from breaking. It prefixes
content rootpath the image sources. 

It excludes the sources that starts with http:// or https:// or /.

If image source starts with './' i.e. image is in the same folder as
content, then it will replace ./ with content directory path and then
adds the rootpath.

Once image source is fixed, it should not break, no matter from where
content is accessed i.e. index or individual page.