Corrected model set for jade rendering
JadeModel implements scopes for the model data. Earlier only the current
model was being passed which caused missing of config model. So changed
it to pass current instance of JadeModel which has get method
implemented to iterate through all scoped models. Earlier implementation
was causing model.get("config") to be null in TagExtractor for
JadeEngine but was working properly for other engines.