Work around a regression introduced in SLF4J 1.7.15
SLF4J version 1.7.15 introduced a regression, where the last formatting
anchor, if it corresponds to a throwable/exception, will not be replaced
with the stringified throwable/exception and the stack trace is
additionally logged. See:
http://mailman.qos.ch/pipermail/slf4j-user/2017-August/001639.html

JBake is using SLF4J 1.7.25, which is affected.

This issue was found by SLF4J Helper for NetBeans IDE:
http://plugins.netbeans.org/plugin/72557/