Fixing issue #42
This small change resolves a `NullPointerException` that occurs when the `JBakeServeTask` is trying to grab the input directory from the Gradle convention mapping. This is because it is trying to directly access the `input` field, rather than calling the accessor method. This switch allows the task to get the correctly configured value.