feat: allow config file to be specified on command line
Adds the `-c VAL` or `--config VAL` command line arguments to allow VAL to be used as a configuration file instead of `jbake.properties`.

Combined with #639, this allows a single source to be used for multiple purposes (for example, the help content of an app could be packaged for a site using one set of templates and for the app using a different set of templates).